requirements.txt
safeconnectors @ git+ssh://git@github.com/hashcib/safeconnectors@92ebbf8828b66afe9cd72a37dc035bc84d56b137
termcolor==1.1.0
websockets==9.1
requests==2.26.0
pandas==1.3.4
pyTelegramBotAPI==4.1.1
elasticsearch==7.15.2
psycopg2-binary==2.9.2
boto3==1.20.23
